+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This is the changelog for Tarmac.
 - Add `python` workflow step type
 - Add `--output-format` argument to command
 - Add `--log-level` argument to command
+- Add `--output-file` argument to command
 
 ### Changed