+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This is the changelog for Tarmac.
 - Add `colored-text` as default output format
 - Add `Failure` exception for use in scripts
 - Add JSON schema for workflow definition files
+- Add `cwd` and `env` arguments to `run` workflow step
 
 ### Changed